ADD rate-limit middlewares and doc
